+++ src/Components/month-switcher/month-switcher.tsx
@@ -1,5 +1,4 @@
 import './monthSwitcher.css';
-import { Component } from 'react';
 import arrowPrev from '../../assets/img/arrow-prev.svg';
 import arrowNext from '../../assets/img/arrow-next.svg';
 
@@ -8,21 +7,20 @@ export interface MonthSwitcherProps{
     monthName: string
 }
 
-class MonthSwitcher extends Component<MonthSwitcherProps> {
 
-    render(){
-        return(
-            <div className="month-switcher">
-                <button onClick = {(e) => this.props.dateChanger(e, -1)}>
-                    <img src={arrowPrev} alt="prev-arrow-icon"/>
-                </button>
-                <span>{this.props.monthName}</span>
-                <button onClick = {(e) => this.props.dateChanger(e, 1)}>
-                    <img src={arrowNext} alt="nex-arrow-icon"/>
-                </button>
-            </div>
-        )
-    }
-} 
+const MonthSwitcher = (props: MonthSwitcherProps) => {
+    const { monthName, dateChanger } = props; 
+    return(
+        <div className="month-switcher">
+            <button onClick = {(e) => props.dateChanger(e, -1)}>
+                <img src={arrowPrev} alt="prev-arrow-icon"/>
+            </button>
+            <span>{monthName}</span>
+            <button onClick = {(e) => dateChanger(e, 1)}>
+                <img src={arrowNext} alt="nex-arrow-icon"/>
+            </button>
+        </div>
+    )
+}
 
 export default MonthSwitcher;
