+++ src/Components/table-layout/table-layout.tsx
@@ -1,19 +1,25 @@
 import React, { FormEvent } from 'react';
 import { Component } from "react";
-import { TableState } from '../../utils/tableState';
 import './tableLayout.css';
 import MonthSwitcher from '../month-switcher';
 import TeamTable from '../team';
 import CalendarFooter from '../calendar-footer';
 import { add, format, setDate, isWeekend } from 'date-fns';
 import CalendarHead from "../calendar-head";
-import { Day } from '../../utils/day';
-import { User } from '../../utils/user';
-import { Team } from '../../utils/team';
+import { Day } from '../../utils/models/day';
+import { Team } from '../../utils/models/team';
+import { Store } from '../../utils/models/store';
 import { connect } from 'react-redux';
-import { isVacation } from '../../utils/services/isVacation';
 import {differenceInCalendarDays, getDaysInMonth, isSameDay, startOfMonth} from 'date-fns';
 import Modal from '../modal';
+
+interface TableState{
+    currentDate:Date;
+    daysInMonth:Day[];
+    modalOpened:boolean;
+    modalState:string;
+}
+
 
 
 class TableLayout extends React.Component<any,TableState>{
@@ -26,11 +32,9 @@ class TableLayout extends React.Component<any,TableState>{
     processModal = (event:any)=>{
         event.preventDefault();
         const form = event.target;
-        //console.log(form.startDate.value,form.endDate.value)
         if(differenceInCalendarDays(new Date(form.endDate.value),new Date(form.startDate.value))<0){
             this.setState({modalState:'error'});
         }
-        //console.log(form.vacType.value);
     }
     componentDidMount(){
         const { currentDate } = this.state;
@@ -38,7 +42,6 @@ class TableLayout extends React.Component<any,TableState>{
         this.setDaysInMonth = this.setDaysInMonth.bind(this);
         this.setState({daysInMonth: this.setDaysInMonth(currentDate)});
     }
-
     protected switchMonth(e: Event, switchDirection: number): void{
         let { currentDate, daysInMonth } = this.state;
         const newDate: Date = add(currentDate , {  months: switchDirection });
@@ -48,10 +51,7 @@ class TableLayout extends React.Component<any,TableState>{
             daysInMonth: daysInMonth
         });
     }
-
-
     private setDaysInMonth(currentDate: Date): Day[]{
-        
         let daysArray: Day[] = [];
         for(let i = 1; i <= getDaysInMonth(currentDate); i++){
             let date: Date = setDate(currentDate, i);
@@ -64,8 +64,6 @@ class TableLayout extends React.Component<any,TableState>{
         }
         return daysArray;
     }
-
-
     render(){
         const { currentDate, daysInMonth } = this.state;
         const { teams, vacations } = this.props;
@@ -97,20 +95,18 @@ class TableLayout extends React.Component<any,TableState>{
                         monthName = {format(currentDate, 'MMMM')}
                         vacations = {vacations}
                     />
-            
                 </table>
                 <Modal 
                     currentDate={isSameDay(this.state.currentDate, new Date)?this.state.currentDate:startOfMonth(this.state.currentDate)} 
                     handleSubmit={this.processModal} close={()=>{this.setState({modalOpened: false , modalState:'loading'})}} 
                     isOpened={this.state.modalOpened} modalState={this.state.modalState}>
                 </Modal>
-
             </>
         )
     }}
 
 
-    const mapStateToProps = (store:any) => {
+    const mapStateToProps = (store: Store) => {
         return {
             vacations: store.vacations,
             teams: store.teams
