--- src/Utils/Day.ts
@@ -1,5 +1,0 @@
-export interface Day{
-    date: Date
-    isWeekend: boolean,
-    dayOfWeek?: string
-}
